scripts/local/requirements.txt
clickhouse-connect==0.5.6
python-dotenv==1.0.0
